PIP implemented Changing channel implemented Play / Pause and rewind / fastforward kind of implemented

+++ app/src/main/java/remote/js/nzse/hda/makeremotesgreatagain/HttpCommandWrapper.java
@@ -59,7 +59,7 @@ public class HttpCommandWrapper {
         execute(param, 500);
     }
 
-    public void play(final int offset) {
+    public void play(final long offset) {
         String param = "timeShiftPlay=" + offset;
         execute(param, 500);
     }
